fix(ConfigFile): Add static_cast for double to WORD conversions (C4244) in IniFile
- Fixed 2 C4244 warnings where double values were implicitly converted to WORD
- Added static_cast<WORD>() for wTrShutterMoveWaitTime assignments
- Ensures type safety and eliminates possible data loss warnings

+++ SRC/ConfigFile/IniFile.cpp
@@ -1341,7 +1341,7 @@ void LoadNanoSpecIni(int iIniFile)
 
 // 2009.09.29 bagus SE --{--
 	if (iIniFile == SYSTEM_SR_INI_FILE) {
-		NanoSysData.SrConfig.wTrShutterMoveWaitTime = lg_TrShutterMoveWaitTime * 1000;
+		NanoSysData.SrConfig.wTrShutterMoveWaitTime = static_cast<WORD>(lg_TrShutterMoveWaitTime * 1000);
 	}
 // 2009.09.29 bagus SE --}--
 
@@ -1405,7 +1405,7 @@ void LoadNanoSpecIni(int iIniFile)
 			NanoSysData.SrConfig.dUvGain					= 0.000;
 			NanoSysData.SrConfig.bTransmittance				= FALSE;
 			NanoSysData.SrConfig.iTrUseLensNo				= 0;	// Turret No:1
-			NanoSysData.SrConfig.wTrShutterMoveWaitTime		= 0.000;
+			NanoSysData.SrConfig.wTrShutterMoveWaitTime = static_cast<WORD>(0.000);
 			NanoSysData.SrConfig.bGantryType				= FALSE;
 			NanoSysData.SrConfig.wGantryNumOfMeasPoint		= 1;
 			NanoSysData.SrConfig.bDistance					= FALSE;
